Linting round 1
