+++ x/bech32ibc/module.go
@@ -4,19 +4,17 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/spf13/cobra"
-
 	abci "github.com/cometbft/cometbft/abci/types"
-
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/palomachain/paloma/x/bech32ibc/client/cli"
 	"github.com/palomachain/paloma/x/bech32ibc/keeper"
 	"github.com/palomachain/paloma/x/bech32ibc/types"
+	"github.com/spf13/cobra"
 )
 
 var (
