+++ x/gravity/handler_test.go
@@ -8,11 +8,10 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
 	"github.com/palomachain/paloma/x/gravity/keeper"
 	"github.com/palomachain/paloma/x/gravity/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 // nolint: exhaustruct
@@ -345,7 +344,7 @@ func TestMsgSendToCosmosOverflow(t *testing.T) {
 		CosmosReceiver: myCosmosAddr.String(),
 		Orchestrator:   "",
 	}
-	// Absoulte max value of 2^256 - 1. Previous versions (v0.43 or v0.44) of cosmos-sdk did not support sdk.Int of this size
+	// Absolute max value of 2^256 - 1. Previous versions (v0.43 or v0.44) of cosmos-sdk did not support sdk.Int of this size
 	maxSend := types.ERC20Token{
 		Amount:   sdk.NewIntFromBigInt(biggestBigInt),
 		Contract: tokenETHAddr2,
