+++ x/gravity/keeper/attestation.go
@@ -6,11 +6,10 @@ import (
 	"strconv"
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	"github.com/cosmos/cosmos-sdk/store/prefix"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	"github.com/cosmos/cosmos-sdk/store/prefix"
 	"github.com/palomachain/paloma/x/gravity/types"
 )
