+++ x/gravity/keeper/batch_test.go
@@ -7,10 +7,9 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/palomachain/paloma/x/gravity/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/palomachain/paloma/x/gravity/types"
 )
 
 // nolint: exhaustruct
