+++ x/gravity/keeper/grpc_query.go
@@ -6,7 +6,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
 	"github.com/palomachain/paloma/x/gravity/types"
 )
