+++ x/gravity/keeper/keeper_batch.go
@@ -1,10 +1,9 @@
 package keeper
 
 import (
+	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	"github.com/cosmos/cosmos-sdk/store/prefix"
 	"github.com/palomachain/paloma/x/gravity/types"
 )
