+++ x/gravity/keeper/keeper_logic_call.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
 	"github.com/palomachain/paloma/x/gravity/types"
 )
