+++ x/gravity/keeper/pool_test.go
@@ -6,10 +6,9 @@ import (
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/palomachain/paloma/x/gravity/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/palomachain/paloma/x/gravity/types"
 )
 
 // Tests that the pool is populated with the created transactions before any batch is created
